pong 5 ball.lua
There is an error with the ball, if we have a look on the line 25 and 26, we see that the code is "self.dy = math.random(2) == 1 and -100 or 100"(25 line) & "self.dx = math.random(-50, 50)"(line 26) and on the line 36 and 37 we have the same code. The error is if we run and play the game, we see that the ball is is going to move up or down on the center and we dont want the ball do this. We need to change the values, (line 25 old self.dy, new self.dx), (line 26 old self.dx, new self.dy) and the same on the other lines(36,37). And this is all.